Add Julie's new emulators (post refactoring)

+++ ravenpy/models/base.py
@@ -73,10 +73,9 @@ class Raven:
         "region_id",
     ]
 
-    # This is just to satisfy mypy, which wants to know about those internal classes defined
+    # This is just to satisfy mypy, which wants to know about this internal class defined
     # by the emulators (which are Raven subclasses)
     Params: Any
-    DerivedParams: Any
 
     def __init__(self, workdir: Union[str, Path] = None):  # , identifier: str = None):
         """Initialize the RAVEN model.
@@ -196,7 +195,9 @@ class Raven:
             with open(fn, "w") as f:
                 self._rv_paths.append(fn)
                 content = rvo.content or rvo.to_rv()
-                assert content.strip(), f"{rvx} has no content!"
+                assert (
+                    content.strip()
+                ), f"{rvx} has no content! (did you forget to use `RV.set_tmpl`?)"
                 f.write(content)
 
     def setup(self, overwrite=False):
@@ -599,7 +600,7 @@ class Raven:
                 for row in reader:
                     for (key, val) in zip(header, row):
                         if "DIAG" in key:
-                            val = float(val)
+                            val = float(val)  # type: ignore
                         out[key].append(val)
 
                 out.pop("")
@@ -676,7 +677,8 @@ class Ostrich(Raven):
         self.write_save_best()
 
         # Create symbolic link to executable
-        os.symlink(self.ostrich_exec, str(self.cmd))
+        if not self.cmd.exists():
+            os.symlink(self.ostrich_exec, str(self.cmd))
 
     def _dump_rv(self):
         """write configuration files to disk."""
@@ -687,7 +689,11 @@ class Ostrich(Raven):
         fn = self.exec_path / "ostIn.txt"
         with open(fn, "w") as f:
             self._rv_paths.append(fn)
-            f.write(self.config.ost.content or self.config.ost.to_rv())
+            content = self.config.ost.content or self.config.ost.to_rv()
+            assert (
+                content.strip()
+            ), "OST has no content! (did you forget to use `RV.set_tmpl`?)"
+            f.write(content)
 
         # OstRandomNumbers.txt
         if self.config.ost.random_numbers_path:
@@ -758,24 +764,31 @@ class Ostrich(Raven):
         p = re.findall(r"(\w+)\s*:\s*([\S]+)", ops[0])
         return OrderedDict((k, float(v)) for k, v in p)
 
-    def ost2raven(self, ops):
+    def ost2raven(self, ostrich_params):
         """Return model parameters.
 
-        Notes
-        -----
-        This method should be subclassed by emulators for which Ostrich has different parameters than the original
-        Raven model.
+        Parameters
+        ----------
+        ostrich_params: dict
+
+        If this method is used in the context of an emulator subclass, it will return
+        an instance of the Params dataclass of the emulator (it will also make sure that the param
+        name conversion from Ostrich to Raven is performed); if not, it will return the list of values.
+
         """
 
         if hasattr(self, "Params"):
-            # We are using an emulator, so it has a `Params` internal class
-            n = len(fields(self.Params))
-            pattern = "par_x{}" if n < 8 else "par_x{:02}"
-            names = [pattern.format(i + 1) for i in range(n)]
-            return self.Params(*[ops[n] for n in names])
+            # We are in an emulator subclass, so a properly typed Params dataclass is available
+            raven_params = {}
+            # Perform Ostrich to Raven param name conversion if needed
+            o2r = getattr(self, "ostrich_to_raven_param_conversion", {})
+            r2o = {r: o for o, r in o2r.items()}
+            for f in fields(self.Params):
+                raven_params[f.name] = ostrich_params[r2o.get(f.name, f.name)]
+            return self.Params(**raven_params)
         else:
             # We are using generic Ostrich
-            return ops.values()
+            return ostrich_params.values()
 
     @property
     def calibrated_params(self):
